Decompose more views away from the main view
Eliminate a few constructs, like the geometry reader, which don't serve a purpose for the UI calculations anymore.
Further fine-tune the padding of the UI in the main view.
Silo ValueOperatorButton and NextButton into their own files. Enable flexibility to display the proper direction of ValueOperatorButton and correct NextButton color scheme.